更新库存导出功能 优化库存子选项显示

+++ ruoyi-admin/src/main/java/com/ruoyi/system/mapper/MaterialChildMapper.java
@@ -13,7 +13,7 @@ import java.util.List;
 public interface MaterialChildMapper 
 {
 
-    public List<MaterialChild> selectMaterialChildByMaterialId(Integer materialId);
+    public List<MaterialChild> selectMaterialChildByMaterialId(Long materialId);
     /**
      * 查询物料
      *
